+++ Commandline/ArgumentType.cs
@@ -1,0 +1,9 @@
+public enum ArgumentType {
+	// Universal types
+	Integer,
+	Boolean,
+	String,
+
+	// Specialised types
+	LogLevel
+}