+++ Logging/MessageSeverity.cs
@@ -1,9 +1,9 @@
 namespace Dialogue.Logging;
 
 public enum MessageSeverity {
-    Debug,
-    Info,
-    Warning,
-    Error,
+	Debug,
+	Info,
+	Warning,
+	Error,
 	Fatal
 }